update
git-svn-id: https://net-server.dit.unitn.it/svn/russo/P4S@1202 36f32d58-a78c-4a8d-a364-f3ecc157e047

+++ src/p4s/core/Alternate.java
@@ -485,7 +485,7 @@ public class Alternate extends AlternateDataStructure implements CDProtocol, EDP
         sender = receiver = null;
         sender = ((Alternate) (im.getSender().getProtocol(pid)));
         receiver = ((Alternate) (node.getProtocol(pid)));
-        int chunktopush = (int) im.getChunks()[0];
+        int chunktopush = im.getChunks()[0];
         if (receiver.getDebug() >= 2) {
             System.out.println(CommonState.getTime() + "\tNode " + node.getID() + " recFINISH_PUSH " + im.getChunkids() + " from " + im.getSender().getID());
         }
@@ -887,7 +887,7 @@ public class Alternate extends AlternateDataStructure implements CDProtocol, EDP
         receiver.remActiveDw(node);
         receiver.addSuccessDownload();
         receiver.addActivePullSuccess();
-        int chunkpulled = (int) im.getChunks()[0];
+        int chunkpulled = im.getChunks()[0];
         receiver.addChunk(chunkpulled, Message.PULL_CYCLE);//add success pull in add chunk
         sender = receiver = null;
         return;
